change index.html to focus on events.
Planning to add the fareharbor embed when we get an event.  OR, given that page-loading times with FH may remain unresolved - will possibly create an image that looks like the fareharbor "badge" for each event, that links to the event sign-up item on fareharbor, or possibly to a new page with more details and photos related to that event.